+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# main
+- Fix leaking event objects on Windows.
+
 # Version 0.1.0
 - Add non-trait `is_read_vectored()` and `is_write_vectored()` functions.
 - Improve documentation.